Cambios de conexion API 20%

+++ WebApiNetCore/BO/Approval.cs
@@ -27,7 +27,7 @@ namespace SAP_Core.BO
         public double MargenDocumento { get; set; }
         public string DocEntry { get; set; }
         public int CantidadAnexo { get; set; }
-        public int MargenGanancia { get; set; }
+        public double MargenGanancia { get; set; }
     }
 
     public class OPriceHistoy
